requirements.txt
langchain==0.3.7
langchain-core==0.3.19
python-dotenv==1.0.1
Flask
langchain-google-genai
